requirements.txt
backports-datetime-fromisoformat==2.0.3
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==45.0.6
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
packaging==25.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
zipp==3.23.0
